Remove swap_cleaner_test.go file and moved tests to pipeline_test.go
