--- integration/swap_cleaner_test.go
@@ -1,65 +1,0 @@
-package snickers_test
-
-import (
-	"io"
-	"os"
-
-	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
-	"github.com/snickers/snickers/core"
-	"github.com/snickers/snickers/db/memory"
-	"github.com/snickers/snickers/types"
-)
-
-func cp(dst, src string) error {
-	s, err := os.Open(src)
-	if err != nil {
-		return err
-	}
-	defer s.Close()
-	d, err := os.Create(dst)
-	if err != nil {
-		return err
-	}
-	if _, err := io.Copy(d, s); err != nil {
-		d.Close()
-		return err
-	}
-	return d.Close()
-}
-
-var _ = Describe("Swap Cleaner", func() {
-	Context("when calling", func() {
-		It("should remove local source and local destination", func() {
-			dbInstance, _ := memory.GetDatabase()
-			dbInstance.ClearDatabase()
-
-			exampleJob := types.Job{
-				ID:               "123",
-				Source:           "http://source.here.mp4",
-				Destination:      "s3://user@pass:/bucket/",
-				Preset:           types.Preset{Name: "presetHere", Container: "mp4"},
-				Status:           types.JobCreated,
-				Details:          "",
-				LocalSource:      "/tmp/123/src/KailuaBeach.mp4",
-				LocalDestination: "/tmp/123/dst/KailuaBeach.webm",
-			}
-
-			dbInstance.StoreJob(exampleJob)
-
-			os.MkdirAll("/tmp/123/src/", 0777)
-			os.MkdirAll("/tmp/123/dst/", 0777)
-
-			cp(exampleJob.LocalSource, "./videos/nyt.mp4")
-			cp(exampleJob.LocalDestination, "./videos/nyt.mp4")
-
-			Expect(exampleJob.LocalSource).To(BeAnExistingFile())
-			Expect(exampleJob.LocalDestination).To(BeAnExistingFile())
-
-			core.CleanSwap(dbInstance, exampleJob.ID)
-
-			Expect(exampleJob.LocalSource).To(Not(BeAnExistingFile()))
-			Expect(exampleJob.LocalDestination).To(Not(BeAnExistingFile()))
-		})
-	})
-})
